# docs/openapi.yaml
openapi: 3.1.0
info:
  title: JPYC x402 Facilitator
  version: "0.1.0"
  description: |
    REST surface of the JPYC x402 facilitator. Spec-conformant with
    `coinbase/x402` v2 (`specs/x402-specification-v2.md`).

    The facilitator is the broadcasting party in the x402 protocol: clients
    sign EIP-3009 `transferWithAuthorization` authorizations off-chain, the
    facilitator verifies and submits them, paying gas.
  license:
    name: MIT
servers:
  - url: https://facilitator.jpyc-service.com
    description: Reference hosted deployment (production, mainnets)
  - url: https://facilitator-staging.jpyc-service.com
    description: Reference hosted deployment (staging, testnets)
  - url: http://localhost:8402
    description: Local dev
paths:
  /supported:
    get:
      summary: List supported (scheme, network) kinds and signers.
      responses:
        "200":
          description: Supported kinds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportedResponse"
  /verify:
    post:
      summary: Verify a payment authorization without broadcasting.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Verification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          description: Malformed request body or unsupported scheme/network.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
  /settle:
    post:
      summary: Verify and broadcast a payment authorization.
      description: |
        Idempotency: if the same `(network, payer, nonce)` triple was settled
        successfully within the in-memory NonceCache TTL (default 10 min), the
        cached transaction hash is returned with `success: true` and **no
        re-broadcast occurs**. Workers also persist the broadcast hash for 72
        hours in the per-chain Durable Object. Identical signed authorizations
        can resume receipt verification after mining or expiry without another
        broadcast. A used nonce alone does not prove payment: a successful
        Workers recovery verifies matching AuthorizationUsed and Transfer
        events in a successful receipt.

        Before a new broadcast, more than 30 seconds must remain on Ethereum
        and Sepolia, or 15 seconds on other supported chains. The optional
        receiptTimeoutMs only bounds receipt waiting; it does not cancel a
        transaction. receipt_pending with a transaction hash means the result
        is still being confirmed. Continue checking the same transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettleRequest"
      responses:
        "200":
          description: |
            Settlement attempted. Returns `success: true` for both fresh
            broadcasts and idempotent cache hits; the latter omits an RPC
            round-trip entirely.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementResponse"
        "429":
          description: Rate limited (per-payer cap exceeded).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementResponse"
        "503":
          description: Facilitator relayer wallet is critically low on native gas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementResponse"
  /settle-status:
    post:
      summary: Look up an existing broadcast without sending a transaction.
      description: |
        Uses the same HMAC authentication as /settle when configured. Workers
        consult Durable Object storage before the process cache. known means
        a broadcast hash is recorded, not that payment has succeeded. A
        receipt must still be verified. A missing record is not proof of no
        broadcast, for example if RPC submission succeeded before storage did.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [network, payer, nonce]
              properties:
                network: { $ref: "#/components/schemas/Caip2" }
                payer: { $ref: "#/components/schemas/Address" }
                nonce: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
      responses:
        "200":
          description: Broadcast lookup result.
          content:
            application/json:
              schema:
                type: object
                required: [known]
                properties:
                  known: { type: boolean }
                  txHash: { type: string }
                  broadcastAt: { type: integer }
                  source: { type: string, enum: [durable, cache] }
                  timeline: { $ref: "#/components/schemas/SettlementTimeline" }
  /health:
    get:
      summary: Liveness probe.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
components:
  schemas:
    Address:
      type: string
      pattern: "^0x[a-fA-F0-9]{40}$"
    Caip2:
      type: string
      example: "eip155:137"
    DecimalString:
      type: string
      pattern: "^[0-9]+$"
    PaymentRequirements:
      type: object
      required: [scheme, network, amount, asset, payTo, maxTimeoutSeconds, extra]
      properties:
        scheme: { type: string, enum: [exact] }
        network: { $ref: "#/components/schemas/Caip2" }
        amount: { $ref: "#/components/schemas/DecimalString" }
        asset: { $ref: "#/components/schemas/Address" }
        payTo: { $ref: "#/components/schemas/Address" }
        maxTimeoutSeconds: { type: integer, minimum: 1 }
        extra:
          type: object
          required: [assetTransferMethod, name, version]
          properties:
            assetTransferMethod: { type: string, enum: [eip3009] }
            name: { type: string, example: "JPY Coin" }
            version: { type: string, example: "1" }
            decimals: { type: integer, minimum: 0 }
            symbol: { type: string }
    Authorization:
      type: object
      required: [from, to, value, validAfter, validBefore, nonce]
      properties:
        from: { $ref: "#/components/schemas/Address" }
        to: { $ref: "#/components/schemas/Address" }
        value: { $ref: "#/components/schemas/DecimalString" }
        validAfter: { $ref: "#/components/schemas/DecimalString" }
        validBefore: { $ref: "#/components/schemas/DecimalString" }
        nonce:
          type: string
          pattern: "^0x[a-fA-F0-9]{64}$"
    PaymentPayload:
      type: object
      required: [x402Version, accepted, payload]
      properties:
        x402Version: { type: integer, enum: [2] }
        resource:
          type: object
          required: [url]
          properties:
            url: { type: string }
            description: { type: string }
            mimeType: { type: string }
        accepted: { $ref: "#/components/schemas/PaymentRequirements" }
        payload:
          type: object
          required: [signature, authorization]
          properties:
            signature:
              type: string
              pattern: "^0x[a-fA-F0-9]{130}$"
            authorization: { $ref: "#/components/schemas/Authorization" }
    VerifyRequest:
      type: object
      required: [x402Version, paymentPayload, paymentRequirements]
      properties:
        x402Version: { type: integer, enum: [2] }
        paymentPayload: { $ref: "#/components/schemas/PaymentPayload" }
        paymentRequirements: { $ref: "#/components/schemas/PaymentRequirements" }
    VerifyResponse:
      type: object
      required: [isValid]
      properties:
        isValid: { type: boolean }
        invalidReason: { type: string }
        payer: { $ref: "#/components/schemas/Address" }
    SettleRequest:
      allOf:
        - $ref: "#/components/schemas/VerifyRequest"
        - type: object
          properties:
            receiptTimeoutMs:
              type: integer
              minimum: 1000
              maximum: 120000
              description: Optional receipt wait budget in milliseconds. Defaults to 120000.
    SettlementTimeline:
      type: object
      description: |
        Optional diagnostic timestamps, in Unix milliseconds. Fields may be
        absent for older broadcasts. broadcastAt records RPC acknowledgement,
        not proof of global mempool visibility. blockTimestamp is chain time;
        differences from server clocks are approximate.
      properties:
        receivedAt: { type: integer }
        verifiedAt: { type: integer }
        queueEnteredAt: { type: integer }
        broadcastStartedAt: { type: integer }
        broadcastAt: { type: integer }
        receiptObservedAt: { type: integer }
        blockTimestamp: { type: integer }
    SettlementResponse:
      type: object
      required: [success, transaction, network]
      properties:
        success: { type: boolean }
        errorReason: { type: string }
        payer: { $ref: "#/components/schemas/Address" }
        transaction: { type: string, description: "Known transaction hash, including pending results; empty if unknown." }
        network: { $ref: "#/components/schemas/Caip2" }
        amount: { $ref: "#/components/schemas/DecimalString" }
        extensions:
          type: object
          properties:
            jpyc.settlementTimeline: { $ref: "#/components/schemas/SettlementTimeline" }
    SupportedKind:
      type: object
      required: [x402Version, scheme, network]
      properties:
        x402Version: { type: integer, enum: [2] }
        scheme: { type: string, enum: [exact] }
        network: { $ref: "#/components/schemas/Caip2" }
    SupportedResponse:
      type: object
      required: [kinds, extensions, signers]
      properties:
        kinds:
          type: array
          items: { $ref: "#/components/schemas/SupportedKind" }
        extensions:
          type: array
          items: { type: string }
        signers:
          type: object
          additionalProperties:
            type: array
            items: { $ref: "#/components/schemas/Address" }